made all pages work, implemented basic grid for projects page
